Refactored /api/projects/
