--- Server/api/projects/handler/getProjects.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const Project = require('../models/Project');
-
-module.exports = (req, res) => {
-  Project.find((err, projects) => {
-    if (err) return console.error(err);
-    res(projects);
-  });
-};
