--- Server/api/projects/handler/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  addProject: require('./addProject'),
-  getProjects: require('./getProjects'),
-  getProject: require('./id/getProject'),
-  updateProject: require('./id/updateProject')
-};
